osd/PrimaryLogPG: cancel in-flight manifest ops on interval changing
We don't want to cache any object_contexts through the interval change.
We actually assert that all currently live references are dead by the
time the flush for the next interval completes.

On detecting a interval changing, let's cancel any in-flight manifest
ops, just as we do for any other cache-tier ops.

Fixes: https://tracker.ceph.com/issues/41514
Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>