osd/PrimaryLogPG: set result=0 for successful writes
This is currently being set for *all* write returns in execute_ctx(), but
that behavior will change soon.  Each of these write methods is a bit
different, so it is tedious to audit, so I'm going with simple: everywhere
we do ++ctx->num_write, we also set result=0, and let the error paths
set errors if/when it is appropriate.

Signed-off-by: Sage Weil <sage@redhat.com>