Project Zipper - Bucketlist
Implement and do the initial conversion to the BucketList API for
Zipper.  This invloves doing portions of the User and Bucket API as
well, so enable BucketList to be used.

This is not the final version; the API needs to be pushed further down
into RGWRados code, but this gets the API layer complete.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>