mgr/dashboard: Use checkbox for multi selection.
Note, the 'multiClick' selection mode does not render a checkbox in the header column, this is only available in 'checkbox' mode, but this mode only allows the selection of a row by clicking the checkbox, it's not possible to select a row by clicking somewhere within it. Because of that the 'multiClick' mode is preferred.

Signed-off-by: Volker Theile <vtheile@suse.com>