cmake: use GNU linker on FreeBSD
The LLVM ld linker does not understand the versioning mapping.
This is a "documented" difference between GNU ld en LLVM ld.

So we fall back to the GNU linker in the binutils package.

Fixing this with a option to cmake is not possible when building
FreeBSD packages. Many of these options are filtered in the
ports make system. And do not make it to cmake being called to
build the build-tree.

Signed-off-by: Willem Jan Withagen <wjw@digiware.nl>