mgr/dashboard: Fixes removal of custom tags during pool edit
The problem was that previously added custom application tags that were
saved during creation or edit, couldn't be deleted while editing the
updated pool. The custom tags weren't shown in the application drop down.

Now they can be removed and are shown in a sorted way in the application
drop down.

Fixes: https://tracker.ceph.com/issues/41747
Signed-off-by: Stephan Müller <smueller@suse.com>