mon/MonitorDBStore: improve get_chunk_tx limits
The old version was horribly inefficient in that it would reencode the
transaction on every iteration.

Instead, estimate the size if we add an item and stop it if looks like it
will go over.  This isn't super precise, but it's close enough, since the
limits are approximate.

Drop the single-use helper since it only makes the code harder to
follow.

Signed-off-by: Sage Weil <sage@redhat.com>