backport-resolve-issue: narrow regular expression
The regular expression used to extract tracker URLs from the PR body
was too generous. When there are two URLs on the same line and the
second is a tracker URl, the regex picked up both of them. For example,
the following string matched the regex:

"https://github.com/ceph/ceph/pull/26538, https://tracker.ceph.com/issues/41452"

This commit fixes the issue.

Signed-off-by: Nathan Cutler <ncutler@suse.com>