osd: use new and improved calc_pg_role()
An acting set might be [0,1,2,1], and we want pgs 1.0s2 and 1.0s4 on
osd.1 to get roles 1 and 3, respectively (instead of 1 and -1).  This
allows the second EC PG on the OSD to have role >= 0.

This probably was unnoticed before, but now it leads to hangs in the
rados/thrash-erasure-code collection because proc_lease() bails out
when is_nonprimary() fails (due to role < 0).

Fixes: https://tracker.ceph.com/issues/43189
Signed-off-by: Sage Weil <sage@redhat.com>