cephadm: do not force container removal on unit stop
The -f arg for 'podman rm' will remove the container without actually
stopping it, leaving the running containers and processes (conmon and
daemon) running and generally making a mess of things.

Doing 'podman rm' without the -f on a running container will complain
that the container is still running.

So, first stop, then remove.

Fixes: 3baad9c37ddeae1045c56ef27ca9a3150b70f330
Signed-off-by: Sage Weil <sage@redhat.com>