mgr/dashboard: Fix npm vulnerabilities
Add npm-force-resolutions to help fix npm vulnerabilities.

When you need to force a package to be updated,
add it to the resolutions lists in package.json and run "npm run fix:audit".

This will update package-lock.json and all next "npm ci" will contain the
specified version of the package.

Signed-off-by: Tiago Melo <tmelo@suse.com>