cmake: let vstart depend on radosgwd
in 4943b601, in cmake, the target of executable "radosgw" is renamed
to "radosgwd",  and the static library of "radosgw_a" was renamed to
"radosgw". this broke the tests which expected radosgw to be available
if "tests" was built.

in this change, both "vstart" and "tests" now depend on "radosgwd"
instead of "radosgw".

Signed-off-by: Kefu Chai <kchai@redhat.com>