rados bench: fix the delayed checking of completed ops
When the submission process is done, rados bench will scan the
completed ops in a sequence manner. However, because the order of
completion may not be sequential, some of the completed ops suffer from
delayed checking, causing the average latency of all ops to be
prolonged. The solution is to reuse the checking mechanism used during
the submission process.

Signed-off-by: Jianshen Liu <jliu120@ucsc.edu>