librbd: switch from fsid to mirror within mirror peer
The mirror_peer_ping will now take the mirror uuid of the remote
peer. The mirror image status for the remote peers will be organized
by the mirror uuid. The mirror uuid was needed for translating
the primary image's peer as well as supporting cleanup in the
future when a peer is removed.

Related: https://tracker.ceph.com/issues/43405
Signed-off-by: Jason Dillaman <dillaman@redhat.com>