mount.ceph: fix incorrect options parsing
The strncmp will only compare the first N bytes, this will be a
problem in some case, for example:

$ mount.ceph :/ /mnt/cephfs -o remount,room_size=10

We assume the "room_size" is a new option, we will never receive
any "room_size=10" data in kernel space, because it will be parsed
as "ro". And another example is if we move the "secret" option up
and before the "secretfile", the "secretfile" will never be correctly
parsed.

Signed-off-by: Xiubo Li <xiubli@redhat.com>