crimson/osd: support 1 lossy connection between a heartbeat peer
1. fix dead loop due to racing of one heartbeat connection:

The Peer (renamed from PeerInfo) attempts to identify if the racing is
happening, this is because the heartbeat peer is not necessarily to be
symmetric. If racing is actually happening, Peer then decides to wait
for another side if it loses, or connect proactively if it wins
(whoami > peer).

2. fix dead loop between hb_front and hb_back connections of the same
peer:

For a reset event from either con_front or con_back, the heartbeat class
should not simply remove the related PeerInfo which contains both of
them. Instead, Peer is improved (complicated) to be a session-alike
class which remembers/tracks individual front/back connectivity. And it
only starts to track from a clean ping_history (same with the original
logic) when both are connected, and stop tracking as soon as either of
them is reset.

3. keep the compatibility with classic heartbeat:

The most important thing is to keep healthy/unhealthy check identical to
the original logic, because it has been working for a long time.  The
less important thing is that the original messenger policy and
Heartbeat::handle_ping() are untouched, so supposingly our heartbeat
component can still talk with classic OSD.

Signed-off-by: Yingxin Cheng <yingxin.cheng@intel.com>